ajuste conversao quantidade

+++ src/user/application/usecases/removerFuncionarioCentro.usecase.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { BadRequestException, Inject, Injectable } from '@nestjs/common';
 import { type IUserRepository } from 'src/user/domain/repositories/IUserRepository';
 
 @Injectable()
@@ -8,7 +8,17 @@ export class RemoverFuncionarioCentroUsecase {
     private readonly userRepository: IUserRepository,
   ) {}
   async execute(userId: string, centerId: string): Promise<string> {
+    const functionario = await this.userRepository.listarPermissoesParaCasl(
+      userId,
+      centerId,
+    );
+    const hasMaster = functionario.filter((item) => item.role === 'MASTER');
+    if (hasMaster.length > 0) {
+      throw new BadRequestException(
+        `Você não tem permissão para deletar um funcionário.`,
+      );
+    }
     await this.userRepository.removerFuncionarioDoCentro(userId, centerId);
-    return Promise.resolve('Centro removido com sucesso');
+    return Promise.resolve('Funcionário removido com sucesso');
   }
 }
